requirements.txt
langchain
langchain-ollama
python-dotenv
langchain-community
streamlit
pymupdf
tiktoken